rename device dir to mappers
