+++ mappers/bluetooth_mapper/Dockerfile
