+++ mappers/bluetooth_mapper/README.md
