+++ mappers/bluetooth_mapper/configuration/config.yaml
