+++ mappers/bluetooth_mapper/configuration/config_map_types.go
