+++ mappers/bluetooth_mapper/deployment.yaml
