+++ mappers/bluetooth_mapper/helper/helper.go
