+++ mappers/bluetooth_mapper/watcher/watcher.go
@@ -26,9 +26,9 @@ import (
 	"github.com/golang/glog"
 	"github.com/paypal/gatt"
 
-	"github.com/kubeedge/kubeedge/device/bluetooth_mapper/action_manager"
-	"github.com/kubeedge/kubeedge/device/bluetooth_mapper/data_converter"
-	"github.com/kubeedge/kubeedge/device/bluetooth_mapper/helper"
+	"github.com/kubeedge/kubeedge/mappers/bluetooth_mapper/action_manager"
+	"github.com/kubeedge/kubeedge/mappers/bluetooth_mapper/data_converter"
+	"github.com/kubeedge/kubeedge/mappers/bluetooth_mapper/helper"
 )
 
 var DeviceConnected = make(chan bool)
