+++ mappers/modbus_mapper/Dockerfile
