+++ mappers/modbus_mapper/src/common.js
