+++ mappers/modbus_mapper/src/constant.js
